+++ src/app/beer_garden/api/http/handlers/v1/request.py
@@ -493,6 +493,14 @@ class RequestListAPI(AuthorizationHandler):
             description: Max seconds to wait for request completion. (-1 = wait forever)
             type: float
             default: -1
+          - name: choice_validation_enabled
+            in: query
+            required: false
+            type: boolean
+            description: |
+              Whether choice validation is enabled for the request. For root requests
+              choice validation is enabled by default. Child requests are disabled by
+              default. This flag overrides the default validation logic
           - name: request
             in: formData
             required: false
@@ -568,6 +576,9 @@ class RequestListAPI(AuthorizationHandler):
         ):
             request_model.requester = self.current_user.username
 
+        choice_validation_enabled = self.get_argument(
+            "choice_validation_enabled", default=None
+        )
         wait_future = None
         if self.get_argument("blocking", default="").lower() == "true":
             wait_future = Future()
@@ -581,7 +592,10 @@ class RequestListAPI(AuthorizationHandler):
                     operation_type="REQUEST_CREATE",
                     model=request_model,
                     model_type="Request",
-                    kwargs={"wait_event": wait_future},
+                    kwargs={
+                        "wait_event": wait_future,
+                        "choice_validation_enabled": choice_validation_enabled,
+                    },
                     target_garden_name=request_model.target_garden,
                     source_garden_name=request_model.source_garden,
                 ),