+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 3.33.0
 TBD
 
+- Added Choice Validation logic supported by Request Creation API calls. Default behavior is to skip
+  choice validation for child requests. (#2113)
 - Updated Request Output to format JSON output files (#2098)
 - Updating NGINX config timeout for receiving Request responses to 10 minutes (#2099)
 - Fixed Request View issue where duplicate requests shown in Children Tree and Status_Updated_At